+++ test/Unit/CleanRegex/Internal/GroupNameValidatorTest.php
@@ -113,9 +113,9 @@ class GroupNameValidatorTest extends TestCase
             [-15, 'Group index must be a non-negative integer, given: -15'],
             [2.23, 'Group index must be an integer or a string, given: double (2.23)'],
             [null, 'Group index must be an integer or a string, given: null'],
-            ["group\n", "Group name must be an alphanumeric string, not starting with a digit, given: 'group\n'"],
-            ["group\t", "Group name must be an alphanumeric string, not starting with a digit, given: 'group\t'"],
-            ["group\r", "Group name must be an alphanumeric string, not starting with a digit, given: 'group\r'"],
+            ["group\n", "Group name must be an alphanumeric string, not starting with a digit, given: 'group\\n'"],
+            ["a\x7f\x7fb", "Group name must be an alphanumeric string, not starting with a digit, given: 'a[DEL\\x7f][DEL\\x7f]b'"],
+            ["a\xc2\xa0b", "Group name must be an alphanumeric string, not starting with a digit, given: 'a[NBSP\\xc2\\xa0]b'"],
         ];
     }
 }